bench_deepspeed/requirements.txt
deepspeed-mii==0.2.3
mpi4py==3.1.5
sentencepiece==0.2.0
